# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hearsay Game API",
    "description": "Authoritative game state, actions, and memory-proof surfaces.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/snapshot": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Snapshot",
        "operationId": "get_snapshot_v1_runs__run_id__snapshot_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/actions": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Take Action",
        "operationId": "take_action_v1_runs__run_id__actions_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/memories": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Get Memory Lineage",
        "operationId": "get_memory_lineage_v1_runs__run_id__memories_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "proposition_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proposition Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryLineageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/memories/recall": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Recall Memories",
        "operationId": "recall_memories_v1_runs__run_id__memories_recall_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryRecallRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryRecallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "format": "uuid",
            "title": "Idempotency Key"
          },
          "verb": {
            "$ref": "#/components/schemas/ActionVerb"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "content": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "idempotency_key",
          "verb"
        ],
        "title": "ActionRequest"
      },
      "ActionResponse": {
        "properties": {
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "consumed_time": {
            "type": "boolean",
            "title": "Consumed Time"
          },
          "snapshot": {
            "$ref": "#/components/schemas/RunSnapshot"
          }
        },
        "type": "object",
        "required": [
          "action_id",
          "consumed_time",
          "snapshot"
        ],
        "title": "ActionResponse"
      },
      "ActionVerb": {
        "type": "string",
        "enum": [
          "move",
          "observe",
          "read_notice_board",
          "talk",
          "promise_help",
          "confront",
          "sleep"
        ],
        "title": "ActionVerb"
      },
      "CreateRunRequest": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Display Name",
            "default": "Newcomer"
          },
          "seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "default": 1729
          }
        },
        "type": "object",
        "title": "CreateRunRequest"
      },
      "CreateRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "snapshot": {
            "$ref": "#/components/schemas/RunSnapshot"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "snapshot"
        ],
        "title": "CreateRunResponse"
      },
      "DialogueState": {
        "properties": {
          "speaker_id": {
            "type": "string",
            "title": "Speaker Id"
          },
          "speaker_name": {
            "type": "string",
            "title": "Speaker Name"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "speaker_id",
          "speaker_name",
          "text"
        ],
        "title": "DialogueState"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LocationState": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "position": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 3,
            "minItems": 3,
            "title": "Position"
          },
          "neighbors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Neighbors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "position"
        ],
        "title": "LocationState"
      },
      "MemoryLineageResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "proposition_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposition Key"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/MemoryVersionState"
            },
            "type": "array",
            "title": "Versions"
          },
          "transmissions": {
            "items": {
              "$ref": "#/components/schemas/TransmissionState"
            },
            "type": "array",
            "title": "Transmissions"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "versions",
          "transmissions"
        ],
        "title": "MemoryLineageResponse"
      },
      "MemoryRecallRequest": {
        "properties": {
          "holder_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Holder Id"
          },
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query"
          },
          "limit": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 8
          }
        },
        "type": "object",
        "required": [
          "holder_id",
          "query"
        ],
        "title": "MemoryRecallRequest"
      },
      "MemoryRecallResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trace Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "holder_id": {
            "type": "string",
            "title": "Holder Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "memories": {
            "items": {
              "$ref": "#/components/schemas/RecalledMemory"
            },
            "type": "array",
            "title": "Memories"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "run_id",
          "holder_id",
          "query",
          "memories"
        ],
        "title": "MemoryRecallResponse"
      },
      "MemoryVersionState": {
        "properties": {
          "belief_id": {
            "type": "string",
            "format": "uuid",
            "title": "Belief Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "proposition_key": {
            "type": "string",
            "title": "Proposition Key"
          },
          "holder_id": {
            "type": "string",
            "title": "Holder Id"
          },
          "narrative_text": {
            "type": "string",
            "title": "Narrative Text"
          },
          "normalized_position": {
            "additionalProperties": true,
            "type": "object",
            "title": "Normalized Position"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "salience": {
            "type": "number",
            "title": "Salience"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "embedding_model_id": {
            "type": "string",
            "title": "Embedding Model Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "belief_id",
          "version",
          "proposition_key",
          "holder_id",
          "narrative_text",
          "normalized_position",
          "confidence",
          "salience",
          "source_kind",
          "source_id",
          "embedding_model_id",
          "active"
        ],
        "title": "MemoryVersionState"
      },
      "NpcState": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "relationship": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": -100.0,
            "title": "Relationship",
            "default": 0
          },
          "speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speech"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "location_id",
          "color"
        ],
        "title": "NpcState"
      },
      "PlayerState": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "traits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Traits"
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "location_id"
        ],
        "title": "PlayerState"
      },
      "PromiseState": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "promisee_id": {
            "type": "string",
            "title": "Promisee Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "deadline_day": {
            "type": "integer",
            "title": "Deadline Day"
          },
          "deadline_phase": {
            "type": "string",
            "enum": [
              "morning",
              "afternoon",
              "evening",
              "night"
            ],
            "title": "Deadline Phase"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "kept",
              "broken"
            ],
            "title": "Status",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "promisee_id",
          "content",
          "deadline_day",
          "deadline_phase"
        ],
        "title": "PromiseState"
      },
      "RecalledMemory": {
        "properties": {
          "belief_id": {
            "type": "string",
            "format": "uuid",
            "title": "Belief Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "proposition_key": {
            "type": "string",
            "title": "Proposition Key"
          },
          "narrative_text": {
            "type": "string",
            "title": "Narrative Text"
          },
          "semantic_similarity": {
            "type": "number",
            "title": "Semantic Similarity"
          },
          "final_score": {
            "type": "number",
            "title": "Final Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "salience": {
            "type": "number",
            "title": "Salience"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          }
        },
        "type": "object",
        "required": [
          "belief_id",
          "version",
          "proposition_key",
          "narrative_text",
          "semantic_similarity",
          "final_score",
          "confidence",
          "salience",
          "source_id"
        ],
        "title": "RecalledMemory"
      },
      "RunSnapshot": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision",
            "default": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed"
            ],
            "title": "Status",
            "default": "active"
          },
          "day": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Day",
            "default": 1
          },
          "phase": {
            "type": "string",
            "enum": [
              "morning",
              "afternoon",
              "evening",
              "night"
            ],
            "title": "Phase",
            "default": "morning"
          },
          "action_count": {
            "type": "integer",
            "maximum": 18.0,
            "minimum": 0.0,
            "title": "Action Count",
            "default": 0
          },
          "world_tick": {
            "type": "integer",
            "minimum": 0.0,
            "title": "World Tick",
            "default": 0
          },
          "weather": {
            "type": "string",
            "enum": [
              "clear",
              "rain"
            ],
            "title": "Weather",
            "default": "clear"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerState"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationState"
            },
            "type": "array",
            "title": "Locations"
          },
          "npcs": {
            "items": {
              "$ref": "#/components/schemas/NpcState"
            },
            "type": "array",
            "title": "Npcs"
          },
          "promises": {
            "items": {
              "$ref": "#/components/schemas/PromiseState"
            },
            "type": "array",
            "title": "Promises"
          },
          "dialogue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DialogueState"
              },
              {
                "type": "null"
              }
            ]
          },
          "recent_events": {
            "items": {
              "$ref": "#/components/schemas/WorldEvent"
            },
            "type": "array",
            "title": "Recent Events"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "seed",
          "player",
          "locations",
          "npcs"
        ],
        "title": "RunSnapshot"
      },
      "TransmissionState": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "proposition_key": {
            "type": "string",
            "title": "Proposition Key"
          },
          "speaker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Id"
          },
          "listener_id": {
            "type": "string",
            "title": "Listener Id"
          },
          "from_belief_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Belief Id"
          },
          "from_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Version"
          },
          "to_belief_id": {
            "type": "string",
            "format": "uuid",
            "title": "To Belief Id"
          },
          "to_version": {
            "type": "integer",
            "title": "To Version"
          },
          "original_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Text"
          },
          "retold_text": {
            "type": "string",
            "title": "Retold Text"
          },
          "mutation_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mutation Note"
          },
          "trust_at_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust At Time"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id",
            "default": "deterministic"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used",
            "default": false
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "inference_attempts": {
            "type": "integer",
            "title": "Inference Attempts",
            "default": 0
          },
          "inference_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inference Latency Ms"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "proposition_key",
          "speaker_id",
          "listener_id",
          "from_belief_id",
          "from_version",
          "to_belief_id",
          "to_version",
          "original_text",
          "retold_text",
          "mutation_note",
          "trust_at_time",
          "model_id"
        ],
        "title": "TransmissionState"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorldEvent": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "visible": {
            "type": "boolean",
            "title": "Visible",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "text"
        ],
        "title": "WorldEvent"
      }
    }
  }
}